Update and Folders (#34)

+++ app/src/main/java/dev/kaua/squash/Security/Login.java
@@ -33,9 +33,9 @@ import java.util.HashMap;
 import java.util.Objects;
 
 import dev.kaua.squash.Activities.MainActivity;
-import dev.kaua.squash.Activities.SignInActivity;
+import dev.kaua.squash.Activities.Auth.SignInActivity;
 import dev.kaua.squash.Activities.SplashActivity;
-import dev.kaua.squash.Activities.ValidateEmailActivity;
+import dev.kaua.squash.Activities.Auth.ValidateEmailActivity;
 import dev.kaua.squash.Data.Account.AccountServices;
 import dev.kaua.squash.Data.Account.AsyncUser_Follow;
 import dev.kaua.squash.Data.Account.DtoAccount;
@@ -488,13 +488,15 @@ public abstract class Login extends SignInActivity{
     }
 
     public static void clearApplicationData(@NonNull Context context) {
-        File cacheDirectory = context.getCacheDir();
-        File applicationDirectory = new File(Objects.requireNonNull(cacheDirectory.getParent()));
-        if (applicationDirectory.exists()) {
-            String[] fileNames = applicationDirectory.list();
-            if(fileNames != null)
-            for (String fileName : fileNames) {
-                if (!fileName.equals("lib")) deleteFile(new File(applicationDirectory, fileName));
+        final File cacheDirectory = context.getCacheDir();
+        if(cacheDirectory != null && cacheDirectory.getParent() != null){
+            final File applicationDirectory = new File(cacheDirectory.getParent());
+            if (applicationDirectory.exists()) {
+                String[] fileNames = applicationDirectory.list();
+                if(fileNames != null)
+                    for (String fileName : fileNames) {
+                        if (!fileName.equals("lib")) deleteFile(new File(applicationDirectory, fileName));
+                    }
             }
         }
     }
@@ -502,7 +504,7 @@ public abstract class Login extends SignInActivity{
         boolean deletedAll = true;
         if (file != null) {
             if (file.isDirectory()) {
-                String[] children = file.list();
+                final String[] children = file.list();
                 if(children != null)
                     for (String child : children) {
                         deletedAll = deleteFile(new File(file, child)) && deletedAll;
